TABLE_ALTERNATE_ROW_COLOR prop
Hängt addNimbusDefaults etwa mit den Problemen in testAlternateXTable
zusammen?

siehe Kommentare von Kleopatra:
        // JW: Hacking around core issue #6882917
        // which is the underlying reason for issue #1180-swingx
        // (SwingX vs Nimbus table striping)

Habe https://bugs.openjdk.java.net/browse/JDK-6882917 gefunden:
A DESCRIPTION OF THE PROBLEM :

in all uis/components/OS supporting I have seen so for, the first row
has normal background color and the second has the alternate color. With
zero-based index counting that is alternate color in every odd row.
Nimbus/DefaultTableCellRenderer do it the wrong way round: start with
alternate in first row and use normal for second. Reason is that it
calculates the "second" row as being the even ... which is ... ehem ...
wrong in zero based counting. For consolation - we at SwingX had it
incorrect in early versions as well <g>

  Fix is easy: in DefaultTableCellRenderer and SynthTableCellRenderer
replace

if row % 2 == 0

with

if row % 2 != 0

// wurde geändert: https://bugs.openjdk.java.net/browse/JDK-2186928