rename method to JXComboBox.isUsingHighlightersForCurrentValue
PENDING JW to KS (kschaefe): review method naming - doesn't sound like
valid English to me (no native speaker of course :-).
Options are to change the property name to
usingHighlightersForCurrentValue
(as we did in JXMonthView after some debate) or stick to getXX.
Thinking about it: maybe then the property should be usesHighlightersXX,
that is third person singular instead of imperative, like in
tracksVerticalViewport of JTable?